Authorization CRUD

+++ src/tasks/tasks.service.ts
@@ -16,13 +16,19 @@ export class TasksService {
     ) { }
 
 
-    getTask(filterDTO: FilterTaskDTO): Promise<TaskEntity[]> {
-        return this.taskRepository.getTasks(filterDTO);
+    getTask(
+        filterDTO: FilterTaskDTO,
+        user: User
+    ): Promise<TaskEntity[]> {
+        return this.taskRepository.getTasks(filterDTO, user);
     }
 
 
-    async getTaskById(id: number): Promise<TaskEntity> {
-        let found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User
+    ): Promise<TaskEntity> {
+        let found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!found) {
             throw new NotFoundException(`Task ${id} not found`);
@@ -30,16 +36,23 @@ export class TasksService {
         return found;
     }
 
-    async deleteTask(id: number): Promise<void> {
-        let rs = await this.taskRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User
+    ): Promise<void> {
+        let rs = await this.taskRepository.delete({id, userId : user.id});
         if (rs.affected === 0) {
             throw new NotFoundException(`Task ${id} not found`)
         }
 
     }
 
-    async updateTaskStatus(id: number, status: TasksStatus): Promise<TaskEntity> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TasksStatus,
+        user: User
+    ): Promise<TaskEntity> {
+        const task = await this.getTaskById(id, user);
         task.status = status
         task.save();
         return task;
